Code cleanup

+++ kernel/common/video/display.h
@@ -3,7 +3,7 @@
 #include "libs/libc/stdbool.h"
 
 typedef struct display_info {
-    uint32_t  addr;
+    uint64_t addr;
     uint16_t width;
     uint16_t height;
     uint16_t pitch;
